extract notebook editor widget
